requirement.txt
streamlit
Pillow
numpy
plotly
ultralytics